Reorganise AircloakApi into separate files
