+++ src/explorer.api/Models/AircloakApi/Helpers.cs
@@ -1,0 +1,21 @@
+using System;
+using System.Linq;
+using System.Text.Json;
+using System.Text.RegularExpressions;
+
+namespace Explorer.Api.AircloakApi.Helpers
+{
+    /// <summary>
+    /// Helper class for converting Json keys from the .NET standard PascalCase to snake_case
+    /// </summary>
+    public class SnakeCaseNamingPolicy : JsonNamingPolicy
+    {
+        public override string ConvertName(string pascalCase)
+        {
+            var fragments = Regex.Matches(pascalCase, "[A-Z]+[a-z]+")
+                .Select(match => match.Value.ToLower());
+            return String.Join("_", fragments);
+        }
+    }
+
+}
